+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2023-12-12 - jvacaris
+### Fixed
+- **[EMM]** Able challenge someone while in matchmaking.
+- *You Won/Lost* message bug.
+
 ## 2023-12-12 - emadriga
 ### Changed
 - Unifying front look & feel for chat/channel rows